[#3] Set push stage inactive
For now I'm trying to fix the push stage but something else is going on in this stage. Maybe what I need is to change the agent to be a dockerfile and give it my custom maven image with different configuration to be able to authenticate with the dockerhub repo.